Moved 'Transform' to its own module
